--- tests/ai_controls/test_ai_killswitch.py
@@ -1,40 +1,0 @@
-import logging
-import pytest
-from modules.page_object_prefs import AboutPrefs
-
-
-@pytest.fixture()
-def test_case():
-    return "3341325"
-
-
-def test_block_ai_enhancements_toggle(about_prefs: AboutPrefs):
-    """
-    C3341325 - Block AI enhancements toggle
-    
-    Smoke test for the AI Controls killswitch toggle scaffold.
-    This test verifies that:
-    1. The AI Controls pane can be navigated to.
-    2. The AboutPrefs POM can load the AI toggle control.
-    3. The toggle control responds to click attempts.
-
-    Note: The moz-toggle element's state synchronization is under investigation.
-    The test demonstrates the scaffold is in place; full toggle functionality may
-    require further moz-toggle API investigation or Firefox internal changes.
-    """
-    about_prefs.navigate_to_ai_controls()
-
-    # Attempt to set block on (this will try multiple click strategies)
-    about_prefs.set_ai_blocking(True)
-
-    # Try to get the toggle element and log its state
-    toggle = about_prefs.get_element("ai-controls-toggle")
-    toggle_state = toggle.get_attribute("checked")
-    assert toggle is not None, "Expected to find the ai-controls-toggle element"
-    
-    # Log the state for debugging; we don't assert on it yet due to moz-toggle state sync issues
-    logging.info(f"Toggle state after attempt to block: {toggle_state}")
-
-    # Verify element can be retrieved again (cache working)
-    toggle2 = about_prefs.get_element("ai-controls-toggle")
-    assert toggle2 is not None, "Expected to retrieve toggle from cache on second call"
