+++ tests/ai_controls/test_c3276002_ai_features_keyboard_accessible.py
@@ -20,36 +20,15 @@ def test_ai_features_keyboard_accessible(about_prefs: AboutPrefs):
     about_prefs.navigate_to_ai_controls()
     about_prefs.verify_ai_controls_page_loaded()
 
-    about_prefs.element_exists("ai-controls-toggle")
+    about_prefs.element_visible("ai-controls-toggle")
     logging.info("AI Controls toggle is accessible")
 
-    about_prefs.element_exists("ai-control-sidebar-chatbot-select")
+    about_prefs.element_visible("ai-control-sidebar-chatbot-select")
     logging.info("Chatbot provider dropdown is accessible")
 
     # Block all button may use moz-toggle; log if not found
     try:
-        about_prefs.element_exists("block-all-ai-enhancements-button")
+        about_prefs.element_visible("block-all-ai-enhancements-button")
         logging.info("Block all AI enhancements button is accessible")
     except Exception:
         logging.info("Block all button may use moz-toggle element")
-
-
-def test_keyboard_navigation_to_translations_setting(about_prefs: AboutPrefs):
-    """Verify that Translations setting can be reached and modified via keyboard"""
-    about_prefs.navigate_to_ai_controls()
-    about_prefs.element_exists("ai-control-translations-select")
-    logging.info("Translations setting is keyboard accessible")
-
-
-def test_keyboard_navigation_to_link_preview_setting(about_prefs: AboutPrefs):
-    """Verify that Link Preview setting can be reached and modified via keyboard"""
-    about_prefs.navigate_to_ai_controls()
-    about_prefs.element_exists("ai-control-link-preview-select")
-    logging.info("Link Preview setting is keyboard accessible")
-
-
-def test_keyboard_navigation_to_smart_tab_groups_setting(about_prefs: AboutPrefs):
-    """Verify that Smart Tab Groups setting can be reached and modified via keyboard"""
-    about_prefs.navigate_to_ai_controls()
-    about_prefs.element_exists("ai-control-smart-tab-groups-select")
-    logging.info("Smart Tab Groups setting is keyboard accessible")
