+++ tests/ai_controls/test_c3310314_features_remain_after_unblock.py
@@ -17,7 +17,7 @@ def test_enabled_features_remain_after_unblock(about_prefs: AboutPrefs):
     """
     about_prefs.navigate_to_ai_controls()
 
-    about_prefs.set_ai_translations("available")
+    about_prefs.set_ai_translations("enabled")
     translations_state_before = about_prefs.get_ai_translations_state()
 
     about_prefs.set_ai_blocking(True)
